pyproject.toml
[tool.poetry]
name = "mlb-alphabet-game"
version = "0.1.0"
description = ""
authors = ["Greg Finley <gregory.finley@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.12"
requests = "^2.28.1"
google-cloud-bigquery = "^3.3.2"
tweepy = "^4.4.0"
pillow = "^8.4.0"
pytz = "^2022.2.1"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
types-requests = "^2.25.0"
types-pytz = "^2021.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
